requirements.txt
fastapi>=0.104.0
uvicorn>=0.24.0
pandas>=2.2.0
scikit-learn>=1.4.0
numpy>=1.26.0
openai>=1.3.0
mistralai>=1.2.0
python-multipart>=0.0.6
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-dotenv>=1.0.0
plotly>=5.17.0
requests>=2.31.0
pydantic>=2.9.0
openpyxl>=3.1.2
